# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CHAT_API",
    "contact": "edwinagemiler17@gmail.com",
    "description": "...",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "variables": {}
    }
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "register user",
        "operationId": "registeruser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/registeruserrequest"
                  },
                  {
                    "example": {
                      "firstname": "Edwin",
                      "lastname": "Amaranto",
                      "email": "edwinagemiler17@gmail.com",
                      "password": "Academlo123@"
                    }
                  }
                ]
              },
              "example": {
                "firstname": "Edwin",
                "lastname": "Amaranto",
                "email": "edwinagemiler17@gmail.com",
                "password": "Academlo123@"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "get all users",
        "operationId": "getallusers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "edwinagemiler17@gmail.com",
                      "password": "Academlo123@"
                    }
                  }
                ]
              },
              "example": {
                "email": "edwinagemiler17@gmail.com",
                "password": "Academlo123@"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "registeruserrequest": {
        "title": "registeruserrequest",
        "required": [
          "firstname",
          "lastname",
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "firstname": "Edwin",
          "lastname": "Amaranto",
          "email": "edwinagemiler17@gmail.com",
          "password": "Academlo123@"
        }
      },
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "edwinagemiler17@gmail.com",
          "password": "Academlo123@"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}